Converting to Swift 3.0

+++ Pod/Classes/FBAnnotationClusterTemplate.swift
@@ -7,9 +7,10 @@
 //
 
 import Foundation
+import UIKit
 
 public enum FBAnnotationClusterDisplayMode {
-	case SolidColor(sideLength: CGFloat, color: UIColor)
+	case SolidColor(sideLength: Float, color: UIColor)
 	case Image(imageName: String)
 }
 
@@ -36,8 +37,8 @@ public struct FBAnnotationClusterTemplate {
 		self.displayMode = displayMode
 	}
 
-	public init (range: Range<Int>?, sideLength: CGFloat) {
-		self.init(range: range, displayMode: .SolidColor(sideLength: sideLength,
+	public init (range: Range<Int>?, sideLength: Float) {
+		self.init(range: range, displayMode: .SolidColor(sideLength: Float(sideLength),
 		                                                 color: UIColor(red:0.11, green:0.70, blue:0.42, alpha:1.00)))
 	}
 }
